Pass user, teams, and favoriteTeam into app
I'm keeping the logs in here because there's still erratic behavior on "filling" and "emptying" user (the process of getting the db content (theoretically) before passing them into router-view as props

This also adds back the route watcher to settings. Turns out it was necessary to unload data for the 'multiple users on a single machine' case, and not just the 'manually entering a different uid into the route' case. *cough*